emergency-response-backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.5.0
python-dotenv==0.19.0
PyJWT==2.3.0
bcrypt==3.2.0
nltk==3.6.3
certifi==2024.6.2
gunicorn==20.1.0 